Bluetooth: Extend Bluetooth blocklist for blocked manufacturer data
Extend Bluetooth blocklist for blocked manufacturer data so that any
filters in the option that matches blocked manufacturer data will be
excluded or rejected. It also filters out blocked manufacturer data
coming from device’s advertisement before sending them to the clients on
the renderer side.

Bug: 1163207
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3781118
Commit-Queue: Jack Hsieh <chengweih@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1030072}

+++ device/bluetooth/test/fake_central.cc
@@ -35,7 +35,7 @@ T ValueOrDefault(Optional&& opt) {
 }
 
 device::BluetoothDevice::ManufacturerDataMap ToManufacturerDataMap(
-    base::flat_map<uint8_t, std::vector<uint8_t>>&& map) {
+    base::flat_map<uint16_t, std::vector<uint8_t>>&& map) {
   return device::BluetoothDevice::ManufacturerDataMap(
       std::make_move_iterator(map.begin()), std::make_move_iterator(map.end()));
 }
